# [None][fix]Make logprobs faster by not copying over everything