@nvxuanyuc 2025-08-14
The current `update_requests_single_beam_single_step` implementation incurs significant performance overhead proportional to the maximum batch size due to the [.tolist()](https://github.com/NVIDIA/TensorRT-LLM/pull/6842/files#diff-76ef6eceb82cb64d4ea6d49fca82d42ac6c942a45ad9fc8d3ff6a9b49e3b466bL912) operation, creating a large CPU workload (“bubble”) that cannot be hidden with overlap scheduling.

E.g. Llama 3.1 8B on single GPU
- Max BS=1 → ~2.05 ms
- Max BS=4 → ~11.5 ms
- Max BS=64 → ~270 ms

Replacing .tolist() with [.item()](https://github.com/NVIDIA/TensorRT-LLM/pull/6842/files#diff-76ef6eceb82cb64d4ea6d49fca82d42ac6c942a45ad9fc8d3ff6a9b49e3b466bR943) logic avoids unnecessary CPU memory allocation and copies, reducing execution time substantially and making it scale with the number of active requests. With a single active request, execution time of `update_requests_single_beam_single_step` drops to ~36 µs.
<img width="1140" height="489" alt="Screenshot 2025-08-14 at 11 46 08 AM" src="https://github.com/user-attachments/assets/9ea3807a-b7c0-4355-be98-693319640363" />
<img width="1440" height="507" alt="Screenshot 2025-08-14 at 11 49 43 AM" src="https://github.com/user-attachments/assets/b7bebec5-e0a1-4170-bd4c-84606f68523a" />